+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- replication column is now resolved by checking the replication_states map
+- added --hide-replication-states column option to v l and r lv
+
 ## [1.25.1] - 2025-04-02
 
 ### Fixed